jwt authentication completed

+++ controllers/api/v1/posts_api.js
@@ -2,6 +2,7 @@ const Post = require("../../../models/post");
 const Comment = require("../../../models/comment");
 
 module.exports.index = async function (req, res) {
+  
   let posts = await Post.find({})
     .sort("-createdAt")
     .populate({
@@ -25,13 +26,19 @@ module.exports.index = async function (req, res) {
 module.exports.destroy = async function (req, res) {
   try {
     const post = await Post.findById(req.params.id);
-    post.remove();
+    
+    if (post.user == req.user.id) {
+      post.remove();
+      await Comment.deleteMany({ post: req.params.id });
 
-    await Comment.deleteMany({ post: req.params.id });
-
-    return res.json(200, {
-      message: `Post ${post.content}  has been deleted`,
-    });
+      return res.json(200, {
+        message: `Post ${post.content}  has been deleted`,
+      });
+    } else {
+      return res.json(401, {
+        message: "You can't delete this post!",
+      });
+    }
   } catch (err) {
     return res.json(500, {
       message: `Internal Server Error`,
